refactor: clean up component dir
