+++ src/components/layout/HelpBar.tsx
