+++ src/components/layout/HelpBarModern.tsx
