+++ src/components/layout/Terminal.tsx
