+++ src/components/ui/RepositoryCard.tsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import { Card } from './Card';
+import { Card } from './ui/Card';
 import { FontAwesomeIcon } from '@fortawesome/react-fontawesome';
 import { faExternalLinkAlt, faCodeFork } from '@fortawesome/free-solid-svg-icons';
 import { RepositoryNode } from '../utils/types';
